Use double instead of int for numbers and finish execution part. TODO: use exceptions instead of cerr, use function ptr for operations to avoid if/else trees

+++ common-core/Circle5/cpp09/ex01/RPN.cpp
@@ -8,12 +8,12 @@ RPN::~RPN() {}
 
 RPN& RPN::operator=(RPN const& rhs)
 {
-	if (this != &rhs)
-		this->_stack = rhs._stack;
-	return *this;
+    if (this != &rhs)
+        this->_stack = rhs._stack;
+    return *this;
 }
 
-std::stack<float>& RPN::getStack()
+std::stack<double>& RPN::getStack()
 {
-	return this->_stack;
+    return this->_stack;
 }
